requirements.txt
moviepy
pillow
numpy
pyaudio
keyboard